requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.5.2
    # via django
certifi==2022.9.24
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
cryptography==38.0.1
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
django==4.1.2
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-debug-toolbar
    #   django-extensions
    #   django-guardian
    #   django-hashid-field
    #   django-htmx
    #   django-simple-bulma
django-allauth==0.51.0
    # via -r requirements.in
django-debug-toolbar==3.7.0
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-guardian==2.4.0
    # via -r requirements.in
django-hashid-field==3.3.7
    # via -r requirements.in
django-htmx==1.12.2
    # via -r requirements.in
django-simple-bulma==2.5.0
    # via -r requirements.in
django-simple-history==3.1.1
    # via -r requirements.in
django-softdelete==0.9.4
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
hashids==1.3.1
    # via django-hashid-field
idna==3.4
    # via requests
libsass==0.21.0
    # via django-simple-bulma
oauthlib==3.2.2
    # via requests-oauthlib
pillow==9.3.0
    # via -r requirements.in
psycopg2-binary==2.9.5
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.6.0
    # via django-allauth
python3-openid==3.2.0
    # via django-allauth
requests==2.28.1
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
sentry-sdk==1.10.1
    # via -r requirements.in
six==1.16.0
    # via
    #   django-softdelete
    #   libsass
sorl-thumbnail==12.9.0
    # via -r requirements.in
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
urllib3==1.26.12
    # via
    #   requests
    #   sentry-sdk
wheel==0.37.1
    # via django-softdelete
whitenoise==6.2.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
